pyproject.toml
[tool.poetry]
name = "mta-realtime"
version = "0.1.0"
description = ""
authors = ["Nolan Conaway <nolanbconaway@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
beautifulsoup4 = "^4.8"
requests = "^2.22"
google = "^2.0"
gtfs-realtime-bindings = "^0.0.6"
protobuf3-to-dict = "^0.1.5"
imgkit = "^1.0"
pendulum = "^2.0"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pydocstyle = "^4.0"
pylint = "^2.3"
pytest = "^5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
